pyproject.toml
[project]
name = "embedding-metadata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.2",
    "wikipedia>=1.4.0",
    "wikipedia-api>=0.14.0",
]
